Refactor: Created CompanyNumberUtils for padNumber function
